cmo-service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cmo-service"
version = "0.1.0"
description = "CMO Service for GPU Management"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "python-dotenv>=0.19.0",
    "drm-core",  # Local dependency
] 